Fix settings CI hygiene
