+++ frontend/src/components/settings/SettingsWorkbenchSections.tsx
@@ -1,4 +1,5 @@
-export { SettingsAlerts, SettingsHero } from "./SettingsWorkbenchHero"
+export { SettingsAlerts } from "./SettingsWorkbenchAlerts"
+export { SettingsHero } from "./SettingsWorkbenchHero"
 export {
   SettingsAccountHealth,
   SettingsMetrics,
